Revamp login flow and remove admin login
